+++ lib/src/widgets/widgets.dart
@@ -1,6 +1,7 @@
 export '../snackbars/snackbars.dart';
 export 'column_with_space.dart';
 export 'conditional_widget.dart';
+export 'gradient_button.dart';
 export 'hide_keyboard.dart';
 export 'image_picker_button.dart';
 export 'row_with_space.dart';